hvs/requirements.txt
faust
cassandra-driver
confluent_kafka
gevent